refactor: Filelocation & fix: navvar and adminbutton
