+++ Project/src/main/java/com/example/model/Address.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.model;
 
 import jakarta.persistence.*;
 import lombok.Getter;
