+++ Project/src/main/java/com/example/model/AdminForm.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.model;
 
 import lombok.Getter;
 import lombok.Setter;
