+++ Project/src/main/java/com/example/model/Cart.java
@@ -10,7 +10,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import com.example.Member;
 import jakarta.persistence.Column;
 
 @Entity
