+++ Project/src/main/java/com/example/repository/AddressRepository.java
@@ -1,6 +1,9 @@
-package com.example;
+package com.example.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.model.Address;
+
 import java.util.List;
 
 public interface AddressRepository extends JpaRepository<Address, Long> {
