+++ Project/src/main/java/com/example/repository/CartRepository.java
@@ -1,7 +1,8 @@
 package com.example.repository;
 
 import com.example.model.Cart;
-import com.example.Member;
+import com.example.model.Member;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
