+++ Project/src/main/java/com/example/repository/MemberRepository.java
@@ -1,6 +1,8 @@
-package com.example;
+package com.example.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.model.Member;
 
 import java.util.Optional;
