+++ Project/src/main/resources/static/css/mypage.css
@@ -1,0 +1,54 @@
+/* 컨테이너 여백 조정 */
+.container {
+    padding-top: 100px;  /* 네비바 높이만큼 상단 여백 추가 */
+    max-width: 1200px;
+    margin: 0 auto;
+    padding-left: 20px;
+    padding-right: 20px;
+}
+
+/* 환영 메시지 스타일 */
+.welcome-message {
+    text-align: center;
+    margin-bottom: 30px;
+}
+
+/* 관리자 버튼 스타일 */
+.admin-button {
+    display: block;
+    width: 200px;
+    margin: 20px auto;
+    padding: 10px;
+    text-align: center;
+    background-color: #333;
+    color: white;
+    text-decoration: none;
+    border-radius: 5px;
+}
+
+/* 메뉴 섹션 스타일 */
+.menu-section {
+    margin-top: 30px;
+}
+
+.menu-list {
+    list-style: none;
+    padding: 0;
+}
+
+.menu-item {
+    margin: 10px 0;
+}
+
+.menu-item a {
+    text-decoration: none;
+    color: #333;
+    font-size: 16px;
+    padding: 10px;
+    display: block;
+    transition: background-color 0.3s;
+}
+
+.menu-item a:hover {
+    background-color: #f5f5f5;
+}
