+++ Project/src/main/resources/static/css/nav.css
@@ -59,10 +59,16 @@ body {
     margin-right: 20px;
 }
 
+.icons a {
+    text-decoration: none;
+    color: inherit;
+}
+
 .icon {
     width: 30px;
     height: 30px;
     transition: transform 0.3s ease;
+    cursor: pointer; /* 마우스 오버시 포인터 커서 표시 */
 }
 
 .icon:hover {
