+++ Project/src/main/resources/templates/home.html
@@ -13,10 +13,10 @@
 <header>
   <div class="navbar">
     <div class="logo">
-      <a href="/test_newhome"><img src="/images/logo.png" alt="Modaz Logo"></a>
+      <a href="/home"><img src="/images/logo.png" alt="Modaz Logo"></a>
     </div>
     <ul class="nav-links">
-      <li><a href="#">HOME</a></li>
+      <li><a href="/home">HOME</a></li>
       <li><a href="#">SHOP</a></li>
       <li><a href="#">PAGE</a></li>
       <li><a href="#">BLOG</a></li>
