+++ Project/src/main/resources/templates/mypage.html
@@ -1,61 +1,54 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>마이페이지</title>
-    <style>
-        .container {
-            width: 80%;
-            margin: 0 auto;
-            padding: 20px;
-        }
-        .welcome-message {
-            font-size: 1.2em;
-            margin-bottom: 20px;
-        }
-        .menu-section {
-            margin-top: 30px;
-        }
-        .menu-list {
-            list-style: none;
-            padding: 0;
-        }
-        .menu-item {
-            margin: 10px 0;
-            padding: 10px;
-            border: 1px solid #ddd;
-            border-radius: 5px;
-        }
-        .menu-item a {
-            text-decoration: none;
-            color: #333;
-        }
-        .menu-item:hover {
-            background-color: #f5f5f5;
-        }
-        .admin-button {
-            background-color: #4CAF50;
-            color: white;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-            text-decoration: none;
-            display: inline-block;
-            margin-top: 20px;
-        }
-        .admin-button:hover {
-            background-color: #45a049;
-        }
-    </style>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>My Page</title>
+    <link href='https://fonts.googleapis.com/css?family=Montserrat' rel='stylesheet'>
+    <link rel="stylesheet" href="/css/mypage.css">
+    <link rel="stylesheet" href="/css/nav.css">
 </head>
 <body>
+    <header>
+        <div class="navbar">
+            <div class="logo">
+                <a href="/home"><img src="/images/logo.png" alt="Modaz Logo"></a>
+            </div>
+            <ul class="nav-links">
+                <li><a href="/home">HOME</a></li>
+                <li><a href="#">SHOP</a></li>
+                <li><a href="#">PAGE</a></li>
+                <li><a href="#">BLOG</a></li>
+                <li><a href="#">CONTACT</a></li>
+            </ul>
+            <div class="icons">
+                <a href="#" class="icon-container">
+                    <img src="/images/search.svg" alt="Search Icon" class="icon">
+                </a>
+                <th:block th:if="${session.loginMember == null}">
+                    <a href="/login" class="icon-container">
+                        <img src="/images/profile.svg" alt="Login Icon" class="icon">
+                    </a>
+                </th:block>
+                <th:block th:if="${session.loginMember != null}">
+                    <a href="/mypage" class="icon-container">
+                        <img src="/images/profile.svg" alt="Profile Icon" class="icon">
+                    </a>
+                    <a href="/logout" class="icon-container">로그아웃</a>
+                </th:block>
+                <a href="/cart" class="icon-container">
+                    <img src="/images/cart.svg" alt="Cart Icon" class="icon">
+                </a>
+            </div>
+        </div>
+    </header>
+
     <div class="container">
         <div class="welcome-message">
-            <span th:if="${member != null}" th:text="${member.name + '님 환영합니다'}"></span>
+            <span th:if="${session.loginMember != null}" th:text="${session.loginMember.name + '님 환영합니다'}"></span>
         </div>
 
-        <div th:if="${member != null && member.admin}">
+        <div th:if="${session.loginMember != null && session.loginMember.email == 'admin@admin'}">
             <a href="/admin" class="admin-button">관리자 페이지</a>
         </div>
