+++ Project/src/main/resources/templates/product-detail.html
@@ -12,18 +12,34 @@
 
 <header>
     <div class="navbar">
-        <div class="logo"><a href="/test_newhome"><img src="/images/logo.png" alt="Modaz Logo"></a></div>
+        <div class="logo">
+            <a href="/home"><img src="/images/logo.png" alt="Modaz Logo"></a>
+        </div>
         <ul class="nav-links">
-            <li><a href="/test_newhome">HOME</a></li>
+            <li><a href="/home">HOME</a></li>
             <li><a href="#">SHOP</a></li>
             <li><a href="#">PAGE</a></li>
             <li><a href="#">BLOG</a></li>
             <li><a href="#">CONTACT</a></li>
         </ul>
         <div class="icons">
-            <a href="#" class="icon-container"><img src="/images/search.svg" alt="Search Icon" class="icon"></a>
-            <a href="/login?redirect=${pageUrl}" class="icon-container"><img src="/images/profile.svg" alt="Profile Icon" class="icon"></a>
-            <a href="#" class="icon-container"><img src="/images/cart.svg" alt="Cart Icon" class="icon"></a>
+            <a href="#" class="icon-container">
+                <img src="/images/search.svg" alt="Search Icon" class="icon">
+            </a>
+            <th:block th:if="${session.loginMember == null}">
+                <a href="/login" class="icon-container">
+                    <img src="/images/profile.svg" alt="Login Icon" class="icon">
+                </a>
+            </th:block>
+            <th:block th:if="${session.loginMember != null}">
+                <a href="/mypage" class="icon-container">
+                    <img src="/images/profile.svg" alt="Profile Icon" class="icon">
+                </a>
+                <a href="/logout" class="icon-container">로그아웃</a>
+            </th:block>
+            <a href="/cart" class="icon-container">
+                <img src="/images/cart.svg" alt="Cart Icon" class="icon">
+            </a>
         </div>
     </div>
 </header>
